Fleet UI: Device user/Host details page layout changing including split out host header and summary card (#28598)
