+++ changes/28137-my-device-page-layout-changes
@@ -1,0 +1,1 @@
+* Updated "My device page" layout
