+++ frontend/pages/hosts/details/_styles.scss
@@ -10,7 +10,6 @@
 
   @include color-contrasted-sections;
   .header {
-    flex: 100%;
     display: flex;
     flex-direction: column;
   }
