+++ frontend/pages/hosts/details/cards/HostHeader/helpers.tsx
