+++ frontend/pages/hosts/details/cards/HostHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HostHeader";
