+++ frontend/pages/hosts/details/cards/HostSummary/_styles.scss
@@ -1,108 +1,19 @@
-.host-summary {
+.host-summary-card {
   display: flex;
-  flex-direction: column;
-  gap: $pad-medium;
+  gap: $pad-medium $pad-xxlarge;
+  padding: $pad-xxlarge;
+  flex-wrap: wrap;
 
-  &__device-status-tag {
-    margin-left: $pad-small;
-    background-color: $ui-warning;
-    padding: $pad-xsmall;
-    font-size: 10px;
-    font-weight: $bold;
-    border-radius: $border-radius;
-
-    &.warning {
-      background-color: $ui-warning;
-    }
-
-    &.error {
-      color: $core-white;
-      background-color: $core-vibrant-red;
-    }
-  }
-
-  &__last-fetched {
-    font-size: $xx-small;
-    color: $core-fleet-black;
-    margin: 0;
-    margin-left: $pad-medium;
-  }
-
-  &__refetch {
+  // Properly vertically aligns host issue icon
+  .host-issue {
     display: flex;
-    margin-left: $pad-medium;
-    margin-right: $pad-small;
-
-    .refetch-btn {
-      font-size: $x-small;
-      height: 38px;
-
-      &:hover {
-        svg {
-          path {
-            fill: $core-vibrant-blue-over;
-          }
-        }
-      }
-    }
-
-    .refetch-spinner {
-      color: $core-vibrant-blue;
-      cursor: default;
-      font-size: $x-small;
-      height: 38px;
-      opacity: 50%;
-      filter: saturate(100%);
-      margin-left: $pad-small;
-
-      .icon {
-        animation: spin 2s linear infinite;
-      }
-
-      @keyframes spin {
-        0% {
-          transform: rotate(0deg);
-          transform-origin: center center;
-        }
-
-        100% {
-          transform: rotate(360deg);
-          transform-origin: center center;
-        }
-      }
-    }
+    gap: $pad-xsmall;
   }
 
-  .card {
-    &__header {
-      display: flex;
-      align-items: center;
-      gap: $pad-xxsmall;
-      max-height: 20px;
-      .premium-icon-tip {
-        position: relative;
-        top: 3px;
-      }
-    }
-    .component__tooltip-wrapper__tip-text {
-      max-width: 326px;
-    }
-
-    display: flex;
-    gap: $pad-medium $pad-xxlarge;
-    padding: $pad-xxlarge;
-    flex-wrap: wrap;
-
-    // Properly vertically aligns host issue icon
-    .host-issue {
-      display: flex;
-      gap: $pad-xsmall;
-    }
-
-    .no-team {
-      color: $ui-fleet-black-50;
-    }
+  .no-team {
+    color: $ui-fleet-black-50;
   }
+
   .data-set dd {
     overflow: initial;
   }
